Rename directory to `EventTileView`
