+++ src/shared-components/event-tiles/TextualEventView/TextualEventView.test.tsx
