+++ src/shared-components/event-tiles/TextualEventView/TextualEventView.tsx
