+++ src/shared-components/event-tiles/TextualEventView/index.ts
